requirements.txt
streamlit
wordcloud
matplotlib
PyMuPDF
python-docx